[Test] add unit tests for ssl models
