+++ cogdl/trainers/gae_trainer.py
@@ -6,6 +6,7 @@ import torch.sparse
 
 from .base_trainer import BaseTrainer
 
+
 class GAETrainer(BaseTrainer):
     def __init__(self, args):
         self.max_epoch = args.max_epoch
@@ -20,6 +21,14 @@ class GAETrainer(BaseTrainer):
     def fit(self, model, data):
         model = model.to(self.device)
         self.num_nodes = data.x.shape[0]
+        adj_mx = (
+            torch.sparse_coo_tensor(
+                data.edge_index, torch.ones(data.edge_index.shape[1]), torch.Size([data.x.shape[0], data.x.shape[0]])
+            )
+            .to_dense()
+            .to(self.device)
+        )
+        data = data.apply(lambda x: x.to(self.device))
 
         print("Training initial embedding...")
         epoch_iter = tqdm(range(self.max_epoch))
@@ -28,7 +37,7 @@ class GAETrainer(BaseTrainer):
         for epoch in epoch_iter:
             model.train()
             optimizer.zero_grad()
-            loss = model.make_loss(data)
+            loss = model.make_loss(data, adj_mx)
             loss.backward()
             optimizer.step()
             epoch_iter.set_description(f"Epoch: {epoch:03d}, Loss: {loss:.4f}")
